BugFixes
pillBoxObject function only worked pillBoxes with single-digit nums.
This bug is fixed now.

Known bugs:
dragging a pill onto another pills in the #pillsContainer div causes the
new pill to appear 'underneath' the old one, but they both sit in the
same pillBox. This bug does not seem to happen in any other container of
pillBoxes